Partial add of logic into the front-end to show validation feedback
For ease of navigation for the user, it is better to implement some of
the back-end renderings in the front-end instead. This way, instead of
doing a post request directly by going to a route, we do a fetch with a
post method. This way, the user does not leave certain pages.